Move file loading Node.js modules into the loading subdirectory
Part of a larger effort to organize the code in the `src` directory.
