+++ make-sync-function
@@ -2,7 +2,7 @@
 
 var fs = require('fs');
 var path = require('path');
-var syncFunctionLoader = require('./src/sync-function-loader.js');
+var syncFunctionLoader = require('./src/loading/sync-function-loader.js');
 
 var errorStatus = 1;
