+++ src/index.js
@@ -3,7 +3,7 @@
 /**
  * The sync-function-loader module. Reads sync functions from files.
  */
-exports.syncFunctionLoader = require('./sync-function-loader.js');
+exports.syncFunctionLoader = require('./loading/sync-function-loader.js');
 
 /**
  * The test-helper module. Provides a number of conveniences to test the behaviour of document definitions.
