+++ src/index.spec.js
@@ -3,8 +3,10 @@ var index = require('./index');
 
 describe('Main package module', function() {
   it('exposes the public API', function() {
-    expect(index.syncFunctionLoader).to.equal(require('./sync-function-loader'));
-    expect(index.testHelper).to.equal(require('./testing/test-helper'));
-    expect(index.validationErrorFormatter).to.equal(require('./testing/validation-error-formatter'));
+    expect(index).to.eql({
+      syncFunctionLoader: require('./loading/sync-function-loader'),
+      testHelper: require('./testing/test-helper'),
+      validationErrorFormatter: require('./testing/validation-error-formatter')
+    });
   });
 });
