+++ src/loading/document-definitions-loader.js
