+++ src/loading/file-fragment-loader.spec.js
@@ -1,5 +1,5 @@
 var expect = require('chai').expect;
-var simpleMock = require('../lib/simple-mock/index.js');
+var simpleMock = require('../../lib/simple-mock/index.js');
 var mockRequire = require('mock-require');
 
 describe('File fragment loader', function() {
