+++ src/loading/sync-function-loader.js
@@ -9,12 +9,12 @@ exports.load = loadFromFile;
 
 var fs = require('fs');
 var path = require('path');
-var indent = require('../lib/indent.js/indent.min.js');
+var indent = require('../../lib/indent.js/indent.min.js');
 var docDefinitionsLoader = require('./document-definitions-loader.js');
 var fileFragmentLoader = require('./file-fragment-loader.js');
 
 function loadFromFile(docDefinitionsFile) {
-  var syncFuncTemplateDir = path.resolve(__dirname, '../templates');
+  var syncFuncTemplateDir = path.resolve(__dirname, '../../templates');
 
   var syncFuncTemplatePath = path.resolve(syncFuncTemplateDir, 'sync-function-template.js');
   var syncFuncTemplate;
